Rewrote the chunk part.

+++ DownloaderNET/LogMessage.cs
@@ -3,6 +3,6 @@
 public class LogMessage
 {
     public String Message { get; set; } = default!;
-    public Int32 Thread { get; set; }
+    public Int64 Thread { get; set; }
     public Exception? Exception { get; set; }
 }
